feat: update chat app
